Don't duplicate DEFAULT section header
If the DEFAULT section header exists but no key/value
pairs exist within the section, crudini will add a
duplicate DEFAULT section header if you use it to set
a new value in the DEFAULT section.

This patch checks for the existence of an empty DEFAULT
section and skips adding a duplicate section header.  We
need to bypass using RawConfigParser for this, as it will
not allow you to check for the existance of the DEFAULT
section.