generate README and man page from --help
DRY the doc update process and generate the man page
and README using a shell script to combine the EXAMPLES
and help2man to convert the lot to man format.

We generate at dist time so as to avoid pushing the
help2man dependency to packagers.
We checkin the README so that it's easily available
and directly usuable by github.

Note when running `make dist` we should ensure that
docs are updated fully before that is run. I.E. make dist
should not result in any pending git changes, but
if that's the case it should be obvious and can be
addressed before release.