don't use separate locking files for synchronization
These would cause deadlock in cases where the system
is stopped in the small window where these files exist.

The mechanism now used is to always lock the file being edited,
and in the default case (without --inplace), we recheck/relock
the file to ensure it hasn't been renamed while taking the initial lock.