maintain the case of parameters output with --get
Note sections are always handled case sensitively by iniparse,
so we key on whether a parameter is specified to --get
as to whether to operate in case sensitive mode.
Note we're still case insensitive when setting,
but are case retentive to the existing case,
of parameters in the ini file.