add support for enforcing name=val format (instead of name = val)
Add --ini-options=nospace option to enforce removal
of spacing around parameter delimiters.
I.e. use 'name=value' rather than 'name = value'.

Note we don't currently support enforcing spacing like 'name = value'
as that is generally not needed, because if a parser supports spaces
it generally supports lack of spaces.  Also we don't support only
adjusting spacing of specified parameters, as again if mixed spacing
is supported, there would be no need for this option.

Note also we don't default to not using spacing, as it's not normal
for ini parsers to not support spacing, and also updates to existing
ini files may introduce churn if we changed the default now.

* crudini.py: Add support for the new --ini-options command line option,
which is initially accepting the 'nospace' option, and can be extended
to support other ini format variants in future.
The 'nospace' implementation is in two parts, for existing and new
parameters. The reformatting of existing parameters is done in
CrudiniInputFilter.readline(), while formatting of new parameters
is supported by wrapping iniparse's OptionLine class in
CrudiniConfigParser(). Note the latter maps using a
space_around_delimiters parameter to ease possible future
use of the configparser interface rather than iniparse.
* EXAMPLES: Add examples for reformatting and updating with nospace.
* README: Auto updated from above.
* TODO: Remove associated todo item.
* tests/test.sh: Add new test.
* tests/nospace-in.ini: Support new test.
* tests/nospace-out.ini: Likewise.

Closes issue #33